#571 :recycle: rename example elements
